Создать тест для Sort и переименовать ShellSortTest -> SortTest
